Login Page
A login page created with basic javascript. Took me about an hour to
make. I just learned about for() loops and arrays in codeacademy and
wanted a challenge to apply these skills. It’s not perfect, e.g. the
functions only work if there is one user in the array. Adding the
functionality to create multiple users is the next step I think.